% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmov-package.R
\docType{package}
\name{gmov-package}
\alias{gmov}
\alias{gmov-package}
\title{gmov: Generative Movement Validation}
\description{
The package provides trajectory-level generative validation diagnostics for
fitted movement models, with a focus on SSF and iSSF workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AurelienNicosiaULaval/gmov}
  \item \url{https://aureliennicosiaulaval.github.io/gmov/}
  \item Report bugs at \url{https://github.com/AurelienNicosiaULaval/gmov/issues}
}

}
\author{
\strong{Maintainer}: Aurélien Nicosia \email{116816597+AurelienNicosiaULaval@users.noreply.github.com}

}
\keyword{internal}
